Update quickstart to 10.0.0 of appindexing library
The try/catch was removed because that exception is no longer being thrown
by the action builder.
Instead, the exception is returned as a part of the task associated with the
start/end calls.

This quickstart does not have an associated codelab or doc. It is just
open sourced on GitHub as a sample for developers to use/play with.
It will be part of the push that goes out during App Indexing V2 launch.

Change-Id: I665cd23ab6879af05e8f2ded44f4964b41efb7e4