Changing tests to work on release artifacts.
We also change the number of reviews since there exists a bug on
recycler views where smoothScrollToPosition leaves a gap on top,
leaving the tests unable to see the element which was added.

One solution would be to use well known techniques in espresso [1],
but this requires us to add keep rules for RecyclerView methods
like `scrollToPosition`, since these are used internally by espresso.
If you feel adding the keep rule for RecyclerViews is a reasonable change
to make to the quickstart app, I would prefer that.

This solution on the other hand makes the reviews list small enough where
the gap still leaves the review on screen.

Reference
[1] : https://developer.android.com/training/testing/espresso/lists.html