Update EmailPasswordActivity.java
When activity start and user is logged in with email verification: true, the UI is not updating correctly (verify button is still showing & verified status: true).
By reloading current user (when activity start) we can fix this issue. If user is not signed in nothing will happen.
It will also make sure that the user account is not deleted from the firebase server each time when user start the activity.